fix(nodejs): update imports for common/ consolidation and suppress orphaned promise rejections
Update imports from ~/utils/, ~/config/, ~/kafka/ to ~/common/utils/,
~/common/config/, ~/common/kafka/ to match the directory consolidation
on master (#66039). Also suppress orphaned promise rejections in all
Promise.race timeout patterns to prevent unhandled rejections when the
timeout wins but the Redis operation later fails.
