+++ nodejs/src/ingestion/pipelines/sessionreplay/config.ts
@@ -8,13 +8,13 @@ import {
     KAFKA_SESSION_RECORDING_SNAPSHOT_ITEM_DLQ,
     KAFKA_SESSION_RECORDING_SNAPSHOT_ITEM_EVENTS,
     KAFKA_SESSION_RECORDING_SNAPSHOT_ITEM_OVERFLOW,
-} from '~/config/kafka-topics'
+} from '~/common/config/kafka-topics'
+import { isDevEnv } from '~/common/utils/env-utils'
 import { INGESTION_DOWNSTREAM_PRODUCER, type IngestionDownstreamProducer } from '~/ingestion/common/producers'
 import {
     INGESTION_SESSIONREPLAY_PRODUCER,
     type IngestionSessionreplayProducer,
 } from '~/ingestion/pipelines/sessionreplay/shared/outputs/producer-config'
-import { isDevEnv } from '~/utils/env-utils'
 
 import { KAFKA_CONSUMER_GROUP_ID as SESSION_RECORDING_DEFAULT_GROUP_ID } from './constants'
