+++ nodejs/src/ingestion/pipelines/sessionreplay/sessions/session-batch-recorder.ts
@@ -1,5 +1,7 @@
 import { v7 as uuidv7 } from 'uuid'
 
+import { logger } from '~/common/utils/logger'
+import { captureException } from '~/common/utils/posthog'
 import { KafkaOffsetManager } from '~/ingestion/pipelines/sessionreplay/kafka/offset-manager'
 import {
     SessionFeatureBlock,
@@ -9,8 +11,6 @@ import { SessionBlockMetadata } from '~/ingestion/pipelines/sessionreplay/shared
 import { SessionMetadataStore } from '~/ingestion/pipelines/sessionreplay/shared/metadata/session-metadata-store'
 import { KeyStore, RecordingEncryptor, SessionKey } from '~/ingestion/pipelines/sessionreplay/shared/types'
 import { MessageWithTeam } from '~/ingestion/pipelines/sessionreplay/teams/types'
-import { logger } from '~/utils/logger'
-import { captureException } from '~/utils/posthog'
 
 import { SessionBatchMetrics } from './metrics'
 import { SessionBatchFileStorage } from './session-batch-file-storage'
